requirements/main.txt
Django==1.10.6
django-model-utils==2.6.1
psycopg2==2.6.2
pycrypto==2.6.1
mysqlclient==1.3.10
python-dotenv==0.6.0
